+++ CHANGELOG.md
@@ -9,9 +9,17 @@
    [#80](https://github.com/aasm/aasm/issues/80)
    thanks to [@ejlangev](https://github.com/ejlangev))
 
-## 3.1.0 (not yet released)
+## 3.9.0 (not yet released)
 
  * deprecated old aasm_* class methods (old-style DSL), in preparation for AASM v4.0.0
+
+## 3.0.26
+
+ * support state.human_name (aliased to state.localized_name) (see [issue #105](https://github.com/aasm/aasm/issues/105)
+
+## 3.0.25
+
+ * initialize the state even if validation is skipped (for ActiveRecord and Mongoid persistence) (see [issue #103](https://github.com/aasm/aasm/issues/103), thanks to [@vfonic](https://github.com/vfonic) and [@aaronklaassen](https://github.com/aaronklaassen))
 
 ## 3.0.24